feat(profile) add order details
