+++ src/pages/form-page.tsx
@@ -1,0 +1,5 @@
+import { UserForm } from "../components/user-form/user-form"
+
+export const FormPage: React.FC = () => {
+    return <UserForm />
+}
