+++ src/pages/page.module.css
@@ -32,16 +32,29 @@
     transform: translate(-50%, -50%)
 }
 
-.navigation {
+.profile {
+    margin: 0;
     display: flex;
+    flex-direction: row;
+    height: 100vh;
+
+}
+.profile_outlet {
+    margin: 016px;
+    width: 50%;
+}
+
+.navigation {
+    /* display: flex;
     flex-direction: column;
-    align-items: flex-start;
+    align-items: flex-start; */
     padding: 0px;
-    position: absolute;
+    width: 25%;
+    /* position: absolute;
     width: 320px;
     height: 192px;
     left: calc(20% - 480px/2);
-    top: 30%;
+    top: 30%; */
     /* box-sizing: border-box; */
     /* top: 0;
     left: 0; */
