--- src/pages/profile -order-page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const ProfileOrderPage: React.FC = () => {
-    return (
-        <div>ProfileOrderPage</div>
-    )
-}
-
-export default ProfileOrderPage
